+++ src/components/Pricing/PricingTable.js
@@ -1,14 +1,9 @@
 import React from "react"
 import { Query } from "react-apollo"
 import cx from "classnames"
-import RadioBtns from "@santiment-network/ui/RadioBtns"
-import Label from "@santiment-network/ui/Label"
-import Icon from "@santiment-network/ui/Icon"
-import Tooltip from "@santiment-network/ui/Tooltip"
-import Panel from "@santiment-network/ui/Panel/Panel"
+import Toggle from "@santiment-network/ui/Toggle"
 import { CURRENT_USER_QUERY } from "../../gql/user"
 import Features from "../Features/Features"
-import PricingDetailsToggle from "./PricingDetailsToggle.js"
 import PlanRestrictBtn from "./PlanRestrictBtn"
 import { PLANS_QUERY } from "../../gql/plans"
 import PLANS from "./prices"
@@ -17,106 +12,52 @@ import {
   getCurrentSheetsSubscription,
   formatPrice,
   getAlternativeBillingPlan,
-  noBasicPlan
+  noBasicPlan,
 } from "../../utils/plans"
 import styles from "./index.module.scss"
 
 const toggleCardDetails = ({ currentTarget }) =>
   currentTarget.classList.toggle(styles.card_opened)
 
-const billingOptions = [
-  {
-    index: "year",
-    content: (
-      <>
-        Bill yearly <Label accent='waterloo'>(save 10%)</Label>
-      </>
-    ),
-  },
-  { index: "month", content: "Bill monthly" },
-]
-
-const DiscountTooltip = () => (
-  <div className={styles.sanTokens}>
-<Tooltip
-          trigger={
-            <div className={styles.tooltipTrigger}>
-              <span>
-                Holding 1000 SAN tokens will result in a 20% discount on all
-                plans
-              </span>
-              <Icon type='question-round' className={styles.icon} />
-            </div>
-          }
-        >
-          <Panel padding className={styles.tooltip}>
-            <p className={styles.text}>
-              <b>Here’s how to claim the SAN holding discount:</b>
-            </p>
-            <ul className={styles.list}>
-              <li className={styles.item}>
-                - Buy 1000 or more SAN tokens (
-                <a
-                  rel='noopener noreferrer'
-                  target='_blank'
-                  href='https://help.santiment.net/en/articles/2542674-how-to-buy-san'
-                >
-                  here's how
-                </a>
-                )
-              </li>
-              <li className={styles.item}>
-                - Log in to SANbase (
-                <a
-                  rel='noopener noreferrer'
-                  target='_blank'
-                  href='https://app.santiment.net/'
-                >
-                  https://app.santiment.net/
-                </a>
-                ). If you don’t have a SANbase account, you can create one with
-                email or MetaMask{" "}
-              </li>
-              <li className={styles.item}>
-                - After logging in to SANbase, head to{" "}
-                <a
-                  rel='noopener noreferrer'
-                  target='_blank'
-                  href='https://app.santiment.net/account'
-                >
-                  Account settings
-                </a>{" "}
-                and connect your account with your MetaMask wallet
-              </li>
-              <li className={styles.item}>- Refresh this page and proceed with your purchase</li>
-              <li className={styles.item}>
-                - Our system checks your SANbase account for 1000+ SAN during the
-                checkout, and automatically applies a 20% discount
-              </li>
-            </ul>
-            <p className={styles.text}>
-              <b>Note:</b> To claim the 20% discount, you just need to hold/HODL
-              enough SAN. The tokens still belong to you - our system simply
-              checks if you have them in your wallet
-            </p>
-          </Panel>
-        </Tooltip>
-        </div>
-)
+const Billing = ({ selected, onClick }) => {
+  const isYearSelected = selected === "year"
+  return (
+    <>
+      <span
+        onClick={() => onClick("month")}
+        className={cx(
+          styles.billing__option,
+          !isYearSelected && styles.billing__option_active
+        )}
+      >
+        Bill monthly
+      </span>
+      <Toggle
+        className={styles.billing__toggle}
+        isActive={isYearSelected}
+        onClick={() => onClick(isYearSelected ? "month" : "year")}
+      />
+      <span
+        className={cx(
+          styles.billing__option,
+          styles.billing__option_year,
+          isYearSelected && styles.billing__option_active
+        )}
+        onClick={() => onClick("year")}
+      >
+        Bill yearly
+        <span className={styles.billing__save}>save 10%!</span>
+      </span>
+    </>
+  )
+}
 
 export default ({ classes = {}, onDialogClose }) => {
   const [billing, setBilling] = React.useState("year")
   return (
     <>
-      {/* <DiscountTooltip /> */}
       <div className={cx(styles.billing, classes.billing)}>
-        <RadioBtns
-          options={billingOptions}
-          defaultSelectedIndex='year'
-          labelOnRight
-          onSelect={res => setBilling(res)}
-          className={styles.bill}
-        />
+        <Billing selected={billing} onClick={setBilling} />
       </div>
       <Query query={CURRENT_USER_QUERY}>
         {({ data: { currentUser } }) => {
